+++ src/main.cpp
@@ -11,7 +11,7 @@ int main(){
     vector<char> vect;
     tg.generateBoundWord(6, vect);
     vector<BoundaryWord> vectBw = tg.getTileWordVect();
-    for(int i =0; i<tg.getTileWordVect().size(); i++){
+    for(int i =0; i<(int)tg.getTileWordVect().size(); i++){
         Tile t(&vectBw[i]);
         cout<<t;
     }
